conduit: add tests for Data.Conduit.Resumable
Test resumable source finalization, too.

The ResumableSink tests come from the conduit-resumablesink package by
Andrew Miller, with minor modifications.

I did a couple "clumsy" things in my tests:

 * Import Data.Conduit, Data.IORef, etc. unqualified.

 * Use partial do patterns instead of shouldBe.

I know this is inconsistent with the existing code, but I left it this way
for a philosophical reason: writing tests should first and foremost
be easy.  It's better to have some tests that produce ugly failure output
than it is to not have those tests at all.  If a test fails later and we
want to know what the failure output was, we can replace partial patterns
with `shouldBe` as part of the troubleshooting process.