refactor lines(Bounded) to use an explicit buffer
I refactored these functions as an exercise--the original versions took
some thinking to understand, and I wondered if use of an explicit buffer
might result in code that is easier to read.  Upon showing it to
@snoyberg, he wrote:

> I agree that your version is clearer. And looking at it again, I don't
> think my version buys us anything in efficiency. I wouldn't be opposed
> to a pull request to change the implementation to what you're
> providing, though it seems like there might be some redundancy in the
> implementation of linesBounded.

Note that the redundancy was due to an optimization (a separate buffer
is only needed for the first line of each awaited Text) and has been
removed here.