Add alignment constraints to sinkStorable
sinkStorable uses the peek function to read a value from the given
memory location but doesn't pay attention to the alignment constraints.
This makes the code to fail under the armhf architecture (thus making
the code non-portable).

Change the sinkStorableHelper function to use a safer version of the
peek function which first copies the data byte-for-byte to a new,
properly aligned location, and then reads the value. The code reverts to
the old implementation in case ALLOW_UNALIGNED_ACCESS is defined.