Marked resourcet as Trustworthy.
In order to be imported from Safe Haskell modules, resourcet needs to be at
least deemed "Trustworthy". In its current state, most modules could be marked
as "Safe", except for resourcet/Control/Monad/Trans/Resource/Internal.hs, which
relies on Control.Monad.Primitive, which is itself not safe.

Instead of trying to make everything "Safe", this patch only marks resourcet as
"Trustworthy", which puts no restriction upon what it can do or import, and
should therefore be a completely harmless change. It only causes GHC to emit a
warning, since most of the files marked as "Trustworhty" are actually deemed
"Safe" by the compiler, which is why this patch also adds
"-fno-warn-trustworthy-safe" to the compilation flags.

Safe Haskell documentation:
https://downloads.haskell.org/~ghc/8.2.2/docs/html/users_guide/safe_haskell.html