Remove noncanonical definitions
This appeases the -Wnoncanonical-monad-instances warning. This warning
exists because a future GHC release may treat noncanonical definitions
as errors.

See also:
  https://gitlab.haskell.org/ghc/ghc/-/wikis/proposal/monad-of-no-return

(Special note: ordinarily, the remedy here would be to remove the
`(*>) = (>>)` definition. But snoyberg/conduit#497 indicates that this
would cause a performance regression, so instead I have inlined `(>>)`.
The important thing is that, if `(>>)` is later defined to equal `(*>)`,
this doesn't result in a loop.)