feat: runResourceT do not discards exception context
Close https://github.com/snoyberg/conduit/issues/521

When exception are meant to be rethrown, in `runResourceT` and
`bracket_`, instead of using `catch` and `throwIO`, we use
`catchNoPropagate` and `rethrowIO` from base >= 4.21 which allows to
keep the exception context and not add any `WhileHandling` context.

For older version of base, we introduce a wrapper with `CPP` which just
keep the previous behavior.